+++ examples/event_loop.luau
@@ -1,6 +1,6 @@
-local winit = require("@luneweb/winit")
-local window = winit.create_window({})
-local webview = winit.create_webview(window, {
+local wry = require("@luneweb/wry")
+local window = wry.create_window({})
+local webview = wry.create_webview(window, {
 	url = "https://tauri.app/",
 })
 
@@ -10,12 +10,12 @@ ipc = webview:ipc_handler(function(msg)
 end)
 
 local handle
-handle = winit.event_loop(window, function(msg)
-	if msg == winit.events.CloseRequested then
+handle = wry.event_loop(window, function(msg)
+	if msg == wry.events.CloseRequested then
 		window:close()
 		ipc.stop()
 		handle.stop()
 	end
 end)
 
-winit.run()
+wry.run()
