+++ examples/multi_window.luau
@@ -1,19 +1,19 @@
-local winit = require("@luneweb/winit")
+local wry = require("@luneweb/wry")
 
-local window1, handle1 = winit.create_window({}), nil
-local window2, handle2 = winit.create_window({
+local window1, handle1 = wry.create_window({}), nil
+local window2, handle2 = wry.create_window({
 	title = "Tauri",
 }), nil
 
-handle1 = winit.event_loop(window1, function(msg)
-	if msg == winit.events.CloseRequested then
+handle1 = wry.event_loop(window1, function(msg)
+	if msg == wry.events.CloseRequested then
 		window1:close()
 		handle1.stop()
 	end
 end)
 
-handle2 = winit.event_loop(window2, function(msg)
-	if msg == winit.events.CloseRequested then
+handle2 = wry.event_loop(window2, function(msg)
+	if msg == wry.events.CloseRequested then
 		window2:close()
 		handle2.stop()
 	end
@@ -25,11 +25,11 @@ local init_script = [[
 	}
 ]]
 
-local webview = winit.create_webview(window2, {
+local webview = wry.create_webview(window2, {
 	url = "https://tauri.app/",
 	init_script = init_script,
 })
 
-winit.run()
+wry.run()
 
 print(webview:evaluate('("b" + "a" + +"a" + "a").toLowerCase()'))
